requirements.txt
Django==2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.0
pkg-resources==0.0.0
pytz==2020.1
PyYAML==5.3.1
render==1.0.0
sqlparse==0.3.1